Move tests into folders and separate out classes
