+++ test/Carter.Tests/DependencyInjection/ScopedServiceDependency.cs
@@ -1,0 +1,12 @@
+namespace Carter.Tests.DependencyInjection
+{
+    public class ScopedServiceDependency
+    {
+        public ScopedServiceDependency(ScopedRequestDependency scopedRequestDependency)
+        {
+            this.Guid = scopedRequestDependency.GetGuid();
+        }
+
+        public string Guid { get; set; }
+    }
+}
