+++ test/Carter.Tests/DependencyInjection/TransientRequestDependency.cs
@@ -1,0 +1,19 @@
+namespace Carter.Tests.DependencyInjection
+{
+    using System;
+
+    public class TransientRequestDependency
+    {
+        private readonly string instanceId;
+
+        public TransientRequestDependency()
+        {
+            this.instanceId = Guid.NewGuid().ToString();
+        }
+
+        public string GetGuid()
+        {
+            return this.instanceId;
+        }
+    }
+}
