+++ test/Carter.Tests/DependencyInjection/TransientServiceDependency.cs
@@ -1,0 +1,12 @@
+namespace Carter.Tests.DependencyInjection
+{
+    public class TransientServiceDependency
+    {
+        public TransientServiceDependency(TransientRequestDependency transientRequestDependency)
+        {
+            this.TheGuid = transientRequestDependency.GetGuid();
+        }
+
+        public string TheGuid { get; }
+    }
+}
