+++ test/Carter.Tests/Modelbinding/DuplicateTestModel.cs
@@ -1,0 +1,6 @@
+namespace Carter.Tests.Modelbinding
+{
+    public class DuplicateTestModel
+    {
+    }
+}
