+++ test/Carter.Tests/Modelbinding/DuplicateTestModelOne.cs
@@ -1,0 +1,8 @@
+namespace Carter.Tests.Modelbinding
+{
+    using FluentValidation;
+
+    public class DuplicateTestModelOne : AbstractValidator<DuplicateTestModel>
+    {
+    }
+}
