+++ test/Carter.Tests/Modelbinding/DuplicateTestModelTwo.cs
@@ -1,0 +1,8 @@
+namespace Carter.Tests.Modelbinding
+{
+    using FluentValidation;
+
+    public class DuplicateTestModelTwo : AbstractValidator<DuplicateTestModel>
+    {
+    }
+}
