+++ test/Carter.Tests/Modelbinding/PathTestModel.cs
@@ -1,0 +1,7 @@
+namespace Carter.Tests.Modelbinding
+{
+    public class PathTestModel
+    {
+        public string Path { get; set; }
+    }
+}
