+++ test/Carter.Tests/Modelbinding/TestModelNoValidator.cs
@@ -1,0 +1,9 @@
+namespace Carter.Tests.Modelbinding
+{
+    public class TestModelNoValidator
+    {
+        public int MyIntProperty { get; set; }
+
+        public string MyStringProperty { get; set; }
+    }
+}
