+++ test/Carter.Tests/Modelbinding/TestModelValidator.cs
@@ -1,0 +1,13 @@
+namespace Carter.Tests.Modelbinding
+{
+    using FluentValidation;
+
+    public class TestModelValidator : AbstractValidator<TestModel>
+    {
+        public TestModelValidator()
+        {
+            this.RuleFor(x => x.MyIntProperty).GreaterThan(0);
+            this.RuleFor(x => x.MyStringProperty).NotEmpty();
+        }
+    }
+}
