+++ test/Carter.Tests/Modules/ExampleModule.cs
@@ -1,0 +1,13 @@
+namespace Carter.Tests.Modules
+{
+    using System.Threading.Tasks;
+
+    public class ExampleModule : CarterModule
+    {
+        public ExampleModule()
+        {
+            this.Post("/foo", ctx => Task.CompletedTask);
+            this.Post("/bar", ctx => Task.CompletedTask);
+        }
+    }
+}
