+++ test/Carter.Tests/Modules/ExampleModuleWithConflictingRoutes.cs
@@ -1,0 +1,13 @@
+namespace Carter.Tests.Modules
+{
+    using System.Threading.Tasks;
+
+    public class ExampleModuleWithConflictingRoutes : CarterModule
+    {
+        public ExampleModuleWithConflictingRoutes()
+        {
+            this.Post("/foo", ctx => Task.CompletedTask);
+            this.Post("/Foo", ctx => Task.CompletedTask);
+        }
+    }
+}
