+++ test/Carter.Tests/StatusCodeHandlers/StatusCodeHandlerTests.cs
@@ -1,9 +1,8 @@
-namespace Carter.Tests
+namespace Carter.Tests.StatusCodeHandlers
 {
     using System.Net.Http;
     using System.Threading.Tasks;
     using Microsoft.AspNetCore.Hosting;
-    using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.TestHost;
     using Xunit;
 
@@ -34,28 +33,6 @@ namespace Carter.Tests
             var response = await this.httpClient.GetAsync("/statushandler");
             var body = await response.Content.ReadAsStringAsync();
             Assert.Equal("Hello World", body);
-        }
-    }
-
-    public class StatusCodeHandlerModule : CarterModule
-    {
-        public StatusCodeHandlerModule()
-        {
-            this.Get("/statushandler", async (req, res, routeData) =>
-            {
-                res.StatusCode = 418;
-                await res.WriteAsync("Hello");
-            });
-        }
-    }
-
-    public class TeapotStatusCodeHandler : IStatusCodeHandler
-    {
-        public bool CanHandle(int statusCode) => statusCode == 418;
-
-        public async Task Handle(HttpContext ctx)
-        {
-            await ctx.Response.WriteAsync(" World");
         }
     }
 }
